refactor: update api endpoints to v1
this commit updates all api endpoints to use the /v1 prefix. this was necessary to version the api and ensure backward compatibility with older versions.

the vite config was also updated to remove the api version from the proxy target. this is because the api version is now included in the endpoint itself.