refactor(api): API 주소에 버전 추가 (#147)
* refactor: update api endpoints to v1

this commit updates all api endpoints to use the /v1 prefix. this was necessary to version the api and ensure backward compatibility with older versions.

the vite config was also updated to remove the api version from the proxy target. this is because the api version is now included in the endpoint itself.

* fix: update refresh token api endpoint

updates the refresh token api endpoint to include the version number.

this ensures that the correct api endpoint is being called when refreshing the access token. there are no breaking changes.

* fix: update refresh token endpoint url

the refresh token endpoint url was incorrect, leading to failed
refresh token requests. this commit updates the url to the
correct value `api/v1/auth/refresh-token`, resolving the issue.
there are no breaking changes.