feat: integrate websocket for real-time updates
this commit integrates websocket connections for
status and friend-related updates. it removes
the previous pubsub implementation and replaces
it with dedicated websocket connections for user
status and friend requests/acceptances.

the changes include:

- removal of `useSocketNotification.ts`.
- introduction of `useStatusSocket.ts` to manage
  user status updates.
- introduction of `useFriendSocket.ts` to handle
  friend request and acceptance notifications and
  data invalidation.
- modification of `socketEvents.ts` to reflect
  the new event structure.
- update the `FriendPage` to use socket update
- add `@tanstack/react-query-devtools` package.

this approach allows for real-time updates and
a more efficient data management, improving the
user experience. there are no breaking changes.