refactor(friend): 친구 신청 및 수락 API 수정 (#180)
* feat: allow multiple statuses for user search

* fix: update status type to array in usersSearch query invalidation

The `status` parameter in the `usersSearch` query needs to accept an array of strings to correctly invalidate the query when creating or accepting friend requests. The previous implementation was passing a single string, which was not the expected type.

This change updates the `status` parameter to be an array, ensuring that the query is properly invalidated and the UI reflects the correct friend request status after a friend request is created or accepted.

There are no breaking changes.

* feat: integrate websocket for real-time updates

this commit integrates websocket connections for
status and friend-related updates. it removes
the previous pubsub implementation and replaces
it with dedicated websocket connections for user
status and friend requests/acceptances.

the changes include:

- removal of `useSocketNotification.ts`.
- introduction of `useStatusSocket.ts` to manage
  user status updates.
- introduction of `useFriendSocket.ts` to handle
  friend request and acceptance notifications and
  data invalidation.
- modification of `socketEvents.ts` to reflect
  the new event structure.
- update the `FriendPage` to use socket update
- add `@tanstack/react-query-devtools` package.

this approach allows for real-time updates and
a more efficient data management, improving the
user experience. there are no breaking changes.