fix(step-navigator): `StepNavigator` Focus 수정 및 Input 컴포넌트에서 엔터가 작동하지 않는 문제 해결 (#183)
* feat: improve ux with keyboard navigation

this commit improves user experience by adding keyboard
navigation to the landing, sign-in, and home pages.

- adds `useRef` and `useEffect` to focus on the
  `DefaultStepNavigator` component on mount. this allows
  users to navigate the options using the tab key.
- adds `onKeyDown` event to the password input in
  `EmailLoginPage` to allow users to submit the form by
  pressing the enter key.

these changes make the application more accessible and
easier to use for users who prefer keyboard navigation.
there are no breaking changes.

* feat: improve homepage accessibility and styling

This commit improves the homepage by:

- Making the navigator focusable using `useRef` and `useEffect`, improving keyboard accessibility.
- Replacing the individual `gunLeft` and `gunRight` styles with a `gun` recipe that accepts a `side` variant. This simplifies the CSS and makes it more maintainable.

* refactor: rename EmailLoginPage to EmailSignInPage

This change renames the EmailLoginPage component and its associated styles to EmailSignInPage. This is done for improved clarity and consistency within the codebase. The old name was a bit ambiguous, and the new name more accurately reflects the component's purpose: handling email-based sign-in. There are no breaking changes associated with this rename.

The component is also updated to handle the "Enter" key press within the email input field, triggering the handleSelect function with an index of 0. This improves user experience by allowing users to submit the email address by pressing "Enter".

* fix: use constant for continue action index

This change addresses the issue where the index `0` was hardcoded in the `onKeyDown` event handler for both the email and password input fields. This makes the code less readable and harder to maintain.

By introducing a constant `CONTINUE_ACTION_INDEX` and assigning it the value `0`, we improve the clarity and maintainability of the code. The code now explicitly indicates that pressing 'Enter' triggers the continue action based on its index, rather than relying on a magic number.

There are no breaking changes.