feat: enable navigation to chatroom from friend list
this change enables navigation to a specific chatroom
when clicking on a friend in the friend list.

previously, clicking on a friend in the list did nothing.
now, it navigates to the `/friend/chatroom` route with the
friend's id as a query parameter. this allows the user to
immediately start a conversation with the selected friend.

the `li` element was also changed to `button` to provide
better semantics for click events, and the surrounding `ul`
was changed to `div` to better represent the layout.