feat: add toast action to open chatroom from notification
This change adds an "Open" action to the chat message toast notification.

This allows users to quickly navigate to the chatroom of the user who sent the message directly from the notification. This improves user experience and provides a more convenient way to engage in conversations.

There are no breaking changes.